Update main.py
Update changes the FCFS handle function code segment to a function of class Scheduler instead of subclass FCFS

since all subclasses were going to have the same exact process of handling a task, the function was made for the parent class instead of each child with a copied function. 

the only difference between schedules would be how the current/next process would be picked, which is handled in the subclass function itself.

This all may differ for the RR schedule which would need a special handling function since the time is shared.